Removed unused test files; renamed server and EODList test files
